# Move libzmq dependent functions out of libflux-core into libzmqutil