--- a/src/broker/module.c
+++ b/src/broker/module.c
@@ -22,6 +22,7 @@
 #include <sys/syscall.h>
 #endif
 
+#include "src/common/libzmqutil/msg_zsock.h"
 #include "src/common/libczmqcontainers/czmq_containers.h"
 #include "src/common/libutil/log.h"
 #include "src/common/libutil/iterators.h"
@@ -253,7 +254,7 @@ flux_msg_t *module_recvmsg (module_t *p)
     int type;
     struct flux_msg_cred cred;
 
-    if (!(msg = flux_msg_recvzsock (p->sock)))
+    if (!(msg = zmqutil_msg_recv (p->sock)))
         goto error;
     if (flux_msg_get_type (msg, &type) < 0)
         goto error;
@@ -314,7 +315,7 @@ int module_sendmsg (module_t *p, const flux_msg_t *msg)
                 goto done;
             if (flux_msg_route_push (cpy, p->modhash->uuid_str) < 0)
                 goto done;
-            if (flux_msg_sendzsock (p->sock, cpy) < 0)
+            if (zmqutil_msg_send (p->sock, cpy) < 0)
                 goto done;
             break;
         }
@@ -323,12 +324,12 @@ int module_sendmsg (module_t *p, const flux_msg_t *msg)
                 goto done;
             if (flux_msg_route_delete_last (cpy) < 0)
                 goto done;
-            if (flux_msg_sendzsock (p->sock, cpy) < 0)
+            if (zmqutil_msg_send (p->sock, cpy) < 0)
                 goto done;
             break;
         }
         default:
-            if (flux_msg_sendzsock (p->sock, msg) < 0)
+            if (zmqutil_msg_send (p->sock, msg) < 0)
                 goto done;
             break;
     }