--- a/src/common/libflux/message.h
+++ b/src/common/libflux/message.h
@@ -130,17 +130,6 @@ int flux_msg_frames (const flux_msg_t *msg);
  */
 flux_msg_t *flux_msg_decode (const void *buf, size_t size);
 
-/* Send message to zeromq socket.
- * Returns 0 on success, -1 on failure with errno set.
- */
-int flux_msg_sendzsock (void *dest, const flux_msg_t *msg);
-int flux_msg_sendzsock_ex (void *dest, const flux_msg_t *msg, bool nonblock);
-
-/* Receive a message from zeromq socket.
- * Returns message on success, NULL on failure with errno set.
- */
-flux_msg_t *flux_msg_recvzsock (void *dest);
-
 /* Get/set message type
  * For FLUX_MSGTYPE_REQUEST: set_type initializes nodeid to FLUX_NODEID_ANY
  * For FLUX_MSGTYPE_RESPONSE: set_type initializes errnum to 0