… doc/man3/flux_reactor_create.rst
… doc/man3/flux_zmq_watcher_create.rst
… doc/man3/index.rst
… src/broker/test/overlay.c
… src/common/libflux/message_iovec.h
… src/common/libflux/message_private.h
… src/common/libflux/message_proto.c
… src/common/libflux/message_proto.h
… src/common/libflux/message_route.c
… src/common/libflux/message_route.h
… src/common/libflux/reactor_private.h
… src/common/libflux/test/message.c
… src/common/libflux/test/reactor.c
… src/common/libjob/Makefile.am
… src/common/libkvs/Makefile.am
… src/common/libpmi/Makefile.am
… src/common/librouter/Makefile.am
… src/common/libterminus/Makefile.am
… src/common/libtestutil/util.c
… src/common/libutil/Makefile.am
… src/common/libzmqutil/Makefile.am
… src/common/libzmqutil/ev_zmq.c
… src/common/libzmqutil/ev_zmq.h
… src/common/libzmqutil/msg_zsock.c
… src/common/libzmqutil/msg_zsock.h
… src/common/libzmqutil/reactor.c
… src/common/libzmqutil/reactor.h
… src/common/libzmqutil/test/ev.c
… src/common/libzmqutil/test/msg_zsock.c
… src/common/libzmqutil/test/reactor.c