@garlick 2021-07-24
Whew!  Nice.

Quick comment on function naming.  What about:

```
flux_msg_sendzsock -> zmqutil_msg_send
flux_msg_recvzsock -> zmqutil_msg_recvz
flux_zmq_watcher_create -> zmqutil_watcher_create
flux_zmq_watcher_get_zsock -> zmqutil_watcher_get_zsock
```
just reducing the redundancy a bit.  Same with the file names - could probably drop the `zmqutil_` prefix since they are already in a zmqutil directory.