@chu11 2021-07-25
re-pushed per comments.  Comments suggested rename of functions were liked, so removed the squashed fixes.