@garlick 2021-07-26
I checked `src/shell/.libs/flux-shell` and it is still linked against libzmq and friends, but I can't for the life of me figure out how!