@chu11 2021-07-26
@garlick I'm not seeing it on a TOSS3 cluster.  Are you ldd-ing `.libs/flux-shell` or `.libs/lt-flux-shell`?  b/c the former will link to the `/usr/lib64/libflux-core.so`, which still has the czmq dependency.