@garlick 2021-07-26
Ah! that was the problem.  I'm not getting an `lt-flux-shell` in my build (Ubuntu 20.04 LTS) but after installing this version over the top of the old one I had installed, no more libzmq!
```
ldd flux-shell
	linux-vdso.so.1 (0x00007fffba7e3000)
	libflux-core.so.2 => /usr/local/lib/libflux-core.so.2 (0x00007f0321102000)
	libpthread.so.0 => /lib/x86_64-linux-gnu/libpthread.so.0 (0x00007f03210df000)
	libflux-optparse.so.1 => /usr/local/lib/libflux-optparse.so.1 (0x00007f03210c9000)
	liblua5.1.so.0 => /lib/x86_64-linux-gnu/liblua5.1.so.0 (0x00007f0321098000)
	libhwloc.so.15 => /lib/x86_64-linux-gnu/libhwloc.so.15 (0x00007f0321047000)
	libjansson.so.4 => /lib/x86_64-linux-gnu/libjansson.so.4 (0x00007f0321038000)
	libm.so.6 => /lib/x86_64-linux-gnu/libm.so.6 (0x00007f0320ee7000)
	libc.so.6 => /lib/x86_64-linux-gnu/libc.so.6 (0x00007f0320cf5000)
	libuuid.so.1 => /lib/x86_64-linux-gnu/libuuid.so.1 (0x00007f0320cec000)
	libdl.so.2 => /lib/x86_64-linux-gnu/libdl.so.2 (0x00007f0320ce6000)
	/lib64/ld-linux-x86-64.so.2 (0x00007f03211ee000)
	libudev.so.1 => /lib/x86_64-linux-gnu/libudev.so.1 (0x00007f0320cb9000)
	libltdl.so.7 => /lib/x86_64-linux-gnu/libltdl.so.7 (0x00007f0320cae000)
```
Yay!