@chu11 2021-07-26
@garlick i think you need to run `src/shell/flux-shell` one time for it to generate the `lt-flux-shell`.