@garlick 2021-07-26
> i think you need to run src/shell/flux-shell one time

I think that's dependent on the libtool version?  My system doesn't produce executables with the `lt-` prefix.  They are the same name as the original.  This is libtool 2.4.6-14.