@grondo 2021-07-26
> I think that's dependent on the libtool version? 

It is dependent on whether "fast-install" mode is enabled in libtool. I guess on some distros fast-install mode isn't needed, so libtool turns it off by default? I could have sworn it was enabled on my system (Ubuntu 18.04), but now I see it is set to `needless`:
```$ grep enable_fast_install config.status
enable_fast_install='needless'
fast_install=$enable_fast_install
```

I did verify that on CentOS 7 (and thus presumably TOSS 3), fast-install is enabled by default:
```bash-4.2$ grep enable_fast_install config.status
enable_fast_install='yes'
fast_install=$enable_fast_install

```