@grondo 2021-07-27
Sorry, I'm not up to date on this PR. However, I was going to ask if it would be useful to add some kind of test that ensures the build (maybe just `flux-shell`) hasn't linked against libcrypto? It would be trivial to add and might catch and accidental regression down the line (though I'm not sure how likely that actually is).

Just thought I throw that idea out there.