@chu11 2021-07-28
> Sorry, I'm not up to date on this PR. However, I was going to ask if it would be useful to add some kind of test that ensures the build (maybe just flux-shell) hasn't linked against libcrypto? It would be trivial to add and might catch and accidental regression down the line (though I'm not sure how likely that actually is).

Good idea.  I just added a simple:

```
if ldd ${FLUX_BUILD_DIR}/src/shell/.libs/lt-flux-shell | grep -q zmq                                                                        
then                                                                                                                                        
    exit 1                                                                                                                                  
fi                                                                                                                                          
exit 0  
```

regression test, until I remembered our conversation above regarding the default of `enable_fast_install`.  I'm not sure the above test is portable.  But presumably checking `ldd src/shell/.libs/flux-shell` isn't portable or safe as well.  hmmm.