@garlick 2021-07-28
> presumably checking ldd src/shell/.libs/flux-shell isn't portable or safe

Would `libtool e ldd src/shell/flux-shell` work here?