@chu11 2021-07-30
re-pushed.  To resolve not exporting iovec functions, I created a `libmessageprivate.la" convenience library within `libflux`.  This will export the iovec functions and a collection of helper functions that are needed.  Unfortunately, this approach did increase the size of message private more than before.  `struct flux_msg` and `struct route_id` and `msg_proto_setup()` also had to be added to message private.

note I added these functions to message private.

```
int msg_route_push (flux_msg_t *msg,
                    const char *id,
                    unsigned int id_len);

int msg_route_append (flux_msg_t *msg,
                      const char *id,
                      unsigned int id_len);

void msg_route_clear (flux_msg_t *msg);

int msg_route_delete_last (flux_msg_t *msg);
```

basically I had to make a call about making `route_id_create()` and `route_id_destroy()` callable in `message_private` OR to hide all of their usage behind "route id helper functions".  I did the latter.