@garlick 2021-07-30
OK, finally getting close to done I think.  The `message_private.c` seems like maybe it should be split up topically.  Without going too crazy with refactoring, I do have a modest set of changes I was playing with on a copy of your branch.  Let me try to clean that up and push it somewhere that you can have a look at.