@garlick 2021-07-30
Well FWIW I moved some things around on my `issue3617_zmqutil_refactor` branch.

The changes are just on top of this branch and do not have proper commit messages since they probably would just be squashed if you decide to include them.  The changes are
* rename `libmessageprivate.la` to `libmessage.la`
* include it in `libflux-internal.la`
* split `libmessage_private.[ch]` into `_iovec`, `_proto`, and `_route` source modules

Let me know what you think @chu11.  I'm open to whatever - we need to get this done and move on.