@chu11 2021-07-30
@garlick Ahhh, I didn't think about putting it into libflux-internal.  Makes sense. I think it works.  Let me yoink your commits and squash am where appropriate and make sure all commits rebuild/pass tests accordingly.