@chu11 2021-07-31
hmmm, one builder failed with.

```
  FAIL: test_plugin.t 100 - flux_plugin_load worked
  #   Failed test 'flux_plugin_load worked'
  #   at test/plugin.c line 419.
  Bail out!  Failed to load test plugin: dlopen: test/.libs/plugin_foo.so: undefined symbol: flux_kvs_commit
  ERROR: test_plugin.t - Bail out! Failed to load test plugin: dlopen: test/.libs/plugin_foo.so: undefined symbol: flux_kvs_commit
```

i'm perplexed, as nothing in the unit tests or `src/common/libflux` uses the kvs.  And it's only one builder.  Going to restart.