@chu11 2021-07-31
re-pushed with a fix for the builder failure described above, just had to add a `libkvs/libkvs.la` dependency addition.